pyproject.toml
[tool.poetry]
name = "nonebot-plugin-resolver2"
version = "1.0.0"
description = "NoneBot2 链接分享解析器插件, 支持的解析(BV号/链接/小程序/卡片),支持平台(b站，抖音，网易云，微博，小红书，youtube，tiktok，twitter...)"
authors = ["fllesser <fllessive@gmail.com>"]
readme = "README.md"
packages = [{include = "nonebot-plugin-resolver2"}]
repository = "https://github.com/fllesser/nonebot-plugin-resolver2"
keywords = ["nonebot", "nonebot2", "resolver", "bilibili", "youtube", "tiktok", "twitter"]

[tool.poetry.dependencies]
python = ">=3.9"
aiohttp = "^3.7"
httpx = ">=0.23"
PyExecJS = "^1.5.1"
bilibili-api-python = ">=16.2.0"
aiofiles = ">=0.8.0"
yt-dlp = ">=2024.11.18"
nonebot2 = ">=2.3.0"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot_plugin_localstore = ">=0.6.0"
nonebot-plugin-apscheduler = ">=0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"